Add second popup type support
There are two types of Honest Respondent popups on Testportal - the first one that was in the original cheat is set when the test is configured to block it after a given number of page blurs. The second type is set when the test is configured to only display a warning. I noticed this and added blocking of the second popup type. I hope that thanks to this people will not rebel, that it does not work ;)